+++ Engine/physics/base_object.h
@@ -33,13 +33,13 @@ class PGE_Phys_Object
 public:
     PGE_Phys_Object();
     virtual ~PGE_Phys_Object();
-    virtual float posX(); //!< Position X
-    virtual float posY(); //!< Position Y
+    virtual double posX(); //!< Position X
+    virtual double posY(); //!< Position Y
 
-    float top();
-    float bottom();
-    float left();
-    float right();
+    double top();
+    double bottom();
+    double left();
+    double right();
 
     void setSize(float w, float h);
     virtual void setPos(long x, long y);
@@ -104,7 +104,7 @@ public:
     };
 
     virtual void update();
-    virtual void render(float x, float y);
+    virtual void render(double x, double y);
 };
 
 bool operator< (const PGE_Phys_Object& lhs, const PGE_Phys_Object& rhs);
